requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
kombu==5.3.7
prompt-toolkit==3.0.43
PyJWT==2.8.0
pyotp==2.9.0
python-decouple==3.8
pytz==2024.1
sqlparse==0.5.0
vine==5.1.0
wcwidth==0.2.13
